Abstract out markers as mixins, to prepare for adding circles and polygons
