+++ addon/helpers/assert-key-exists.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+function assertKeyExists(arrayOfObjects, name) {
+  arrayOfObjects.forEach((object) => {
+    Ember.assert('A `key` attribute must be provided for all `google-maps-addon` ' + name, !!object.key);
+  });
+}
+
+export default assertKeyExists;
