+++ addon/mouse-events.js
@@ -1,0 +1,17 @@
+/**
+@let mouseEvents
+@type array
+Supported mouse events by googlemaps
+  ref 'https://developers.google.com/maps/documentation/javascript/reference#events_50'
+**/
+export default [
+  'click',
+  'dblclick',
+  'drag',
+  'dragend',
+  'dragstart',
+  'mousemove',
+  'mouseout',
+  'mouseover',
+  'rightclick'
+];
