+++ addon/objects/polygons.js
@@ -1,0 +1,4 @@
+import Ember from 'ember';
+
+export default Ember.Mixin.create({
+});
